view calendar remove useEffect create new custom hook that implement useMemo for extract data from events
